+++ wp_mobile/src/collection/post_metadata_collection.rs
@@ -127,7 +127,7 @@ impl PostMetadataCollectionWithEditContext {
             Vec::new()
         } else {
             self.service
-                .read_posts_by_ids_from_db(&all_ids)
+                .read_post_full_entities_by_ids_from_db(&all_ids)
                 .map_err(|e| CollectionError::DatabaseError {
                     err_message: e.to_string(),
                 })?
@@ -359,7 +359,7 @@ impl PostMetadataCollectionWithEditContext {
 
         let changed_post = self
             .service
-            .read_posts_by_ids_from_db(&[post_id])
+            .read_post_full_entities_by_ids_from_db(&[post_id])
             .map_err(|e| CollectionError::DatabaseError {
                 err_message: e.to_string(),
             })?
@@ -430,7 +430,7 @@ impl PostMetadataCollectionWithEditContext {
             .collect();
         let all_posts = self
             .service
-            .read_posts_by_ids_from_db(&all_post_ids)
+            .read_post_full_entities_by_ids_from_db(&all_post_ids)
             .map_err(|e| CollectionError::DatabaseError {
                 err_message: e.to_string(),
             })?;